+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## devel
+- Allows fitting the same model to multiple samples
 - Two-Compartments with intrinsic clearance and metabolites added
 
 ## v17.05 Chlorine 32