+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Allows fitting the same model to multiple samples
 - Two-Compartments with intrinsic clearance and metabolites added
 - Two-Compartments with autoinduction added
+- Allometric scaling added
 
 ## v17.05 Chlorine 32