+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - Allows fitting the same model to multiple samples
 - Two-Compartments with intrinsic clearance and metabolites added
 - Two-Compartments with autoinduction added
-- Allometric scaling added
+- Allometric scaling can now be estimated and applied
 - Labels can be rewritten and several labels added at the same time
 
 ## v17.05 Chlorine 32